CY-1732 Functions evaluation: pass a single storage object (#360)
* Refactor Storage usage in tests

* DSL functions evaluation: use a single Storage instead of methods

Instead of passing in all the methods separately, use a storage class
with them.

* dsl functions: get rid of the "plan/storage" + "handler" layer split

Functions are now evaluated based on a single "handler" type, which
has methods like `get_node` etc, which can be based on either a
plan or a runtime storage, but the functions don't care
about which is it.

* Drop unused cases from _args_to_str_func

* Update recurionlimit's docstrings

* DSL functions: remove the evaluate/evaluate_runtime split

Functions are now just evaluated, they themselves don't care if
it's runtime or not. They will get the data they need from the
"handler" which contains either the plan (at creation time) or
db storage (at runtime), but this is abstracted away from the
intrinsic functions themselves behind a single interface.